// api/src/schemas/auth.ts
import { z } from "zod";

export const registerSchema = z.object({
    email: z.string().email("Invalid email address").max(255),
    username: z.string().min(2, "Username must be at least 2 characters").max(100),
    password: z.string().min(8, "Password must be at least 8 characters").max(128, "Password must be at most 128 charaters"),
});

export const loginSchema = z.object({
    email: z.string().email("Invalid email address"),
    password: z.string().min(1, "Password is required"),
});

export type RegisterInput = z.infer<typeof registerSchema>;
export type LoginInput = z.infer<typeof loginSchema>;